requirements.txt
streamlit
pandas
python-docx
st-gsheets-connection